requirements.txt
streamlit>=1.20
openai>=1.0.0
python-dotenv>=1.0.0